fix(ci): dedupe the split-pane chat harness page wiring
capture-chatpane-queue-edit.mjs and capture-chatpane-upload-error.mjs
carried the same 14-line preparePage stanza (route fallback, websocket
silence, split-layout seed), which tips repo-wide jscpd duplication over
the 0% gate on every merge ref. Extract it into
scripts/lib/prepare-split-chat-page.mjs with a `pre` hook for the
upload harness's 400 intercept — the same shape as the serveDist dedupe.
